AppleDouble support

+++ sit.c
@@ -61,13 +61,20 @@
 #include <unistd.h>
 #include <dirent.h>
 #include <libgen.h>
+#include <time.h>
+#include <stdint.h>
+#include <limits.h>
+#include <errno.h>
 #ifdef BSD
 #include <sys/time.h>
-#include <sys/timeb.h>
-#else
-#include <time.h>
 #endif
 #include "sit.h"
+#include "appledouble.h"
+
+/* Type compatibility for non-BSD systems */
+#ifndef ushort
+typedef unsigned short ushort;
+#endif
 
 #ifndef min
 #define min(a,b) ((a)<(b)?(a):(b))
@@ -75,6 +82,41 @@
 
 /* Mac time of 00:00:00 GMT, Jan 1, 1970 */
 #define TIMEDIFF 0x7c25b080
+
+/* Platform compatibility macros */
+#ifdef __APPLE__
+#define HAVE_BIRTHTIME 1
+#define HAVE_NAMEDFORK 1
+#endif
+
+/* Get timezone offset portably */
+static long get_timezone_offset(void) {
+#if defined(BSD) || defined(__GLIBC__)
+    time_t now;
+    struct tm *tm;
+    time(&now);
+    tm = localtime(&now);
+    return tm->tm_gmtoff + (tm->tm_isdst ? 3600 : 0);
+#else
+    /* Fallback: assume UTC */
+    return 0;
+#endif
+}
+
+/* Safe write that checks for errors and partial writes */
+static int safe_write(int fd, const void *buf, size_t count, const char *context) {
+    ssize_t written = write(fd, buf, count);
+    if (written < 0) {
+        fprintf(stderr, "Error writing %s: %s\n", context, strerror(errno));
+        return -1;
+    }
+    if ((size_t)written != count) {
+        fprintf(stderr, "Partial write for %s: wrote %zd of %zu bytes\n",
+                context, written, count);
+        return -1;
+    }
+    return 0;
+}
 
 struct sitHdr sh;
 struct fileHdr fh;
@@ -90,7 +132,7 @@ int verbose;
 char *Creator, *Type;
 
 static void usage(char *arg0) {
-    fprintf(stderr, "Usage: %s ", arg0);
+	fprintf(stderr,"Usage: %s ", arg0);
     fprintf(stderr, "[-v] [-u] [-T type] [-C creator] [-o dstfile] file ...\n");
     fprintf(stderr, "Options:\n");
     fprintf(stderr, "  -v           Verbose output (can specify more than once for extra info)\n");
@@ -104,7 +146,7 @@ static void usage(char *arg0) {
     fprintf(stderr, "  %s file1 file2 file3\n", arg0);
     fprintf(stderr, "  # create \"FolderArchive.sit\" containing FolderToBeArchived\n");
     fprintf(stderr, "  %s -o FolderArchive.sit FolderToBeArchived\n", arg0);
-    fprintf(stderr, "  # specify that untyped files are JPEG and open in GraphicConverter\n");
+fprintf(stderr, "  # specify that untyped files are JPEG and open in GraphicConverter\n");
     fprintf(stderr, "  %s -o jpgArchive.sit -T JPEG -C GKON *.jpg\n", arg0);
 }
 extern char *optarg;
@@ -117,11 +159,10 @@ int put_folder(char *name, int level);
 int put_folder_entry(char *name, long startPos, int mtype, int level);
 int put_file(char *name, int level);
 int dofork(char *name, int convert);
-void cp2(unsigned int x, char *dest);
-void cp4(unsigned long x, char *dest);
+void cp2(uint16_t x, char *dest);
+void cp4(uint32_t x, char *dest);
 
-int main(int argc, char** argv)
-{
+int main(int argc, char **argv) {
 	int i,n;
 	int total=0, items=0;
 	int c;
@@ -167,7 +208,9 @@ int main(int argc, char** argv)
 		exit(1);
 	}
 	/* empty header, will seek back and fill in later */
-	write(ofd,&sh,sizeof sh);
+	if (safe_write(ofd, &sh, sizeof sh, "archive header") < 0) {
+		exit(1);
+	}
 
 	for (i=optind; i<argc; i++) {
 		n = put_item(argv[i]);
@@ -186,15 +229,19 @@ int main(int argc, char** argv)
     strncpy((char*)sh.sig2,"rLau",4);
     sh.version = 1;
 
-	write(ofd,&sh,sizeof sh);
-	close(ofd);
+	if (safe_write(ofd, &sh, sizeof sh, "final archive header") < 0) {
+		exit(1);
+	}
+	if (close(ofd) < 0) {
+		fprintf(stderr, "Error closing archive: %s\n", strerror(errno));
+		exit(1);
+	}
 	if (verbose) {
 		fprintf(stdout, "Wrote %d bytes to \"%s\"\n",total,defoutfile);
 	}
 }
 
-int put_item(char *name)
-{
+int put_item(char *name) {
 	int n = 0;
 	struct stat st;
 	if (lstat(name,&st)==0 && S_ISDIR(st.st_mode)) {
@@ -212,23 +259,35 @@ int put_item(char *name)
 	return n;
 }
 
-int put_folder(char *name, int level)
-{
+int put_folder(char *name, int level) {
 	DIR *dir;
 	struct dirent *entry;
-	char path[BUFSIZ];
+	char path[PATH_MAX];
 	int i, n = 0;
 
 	if (!(dir = opendir(name))) {
 		return 0;
 	}
 	while ((entry = readdir(dir)) != NULL) {
-		snprintf(path, sizeof(path), "%s/%s", name, entry->d_name);
-		if (entry->d_type == DT_DIR) { /* if it's a directory */
+		struct stat entry_st;
+
+		/* Skip . and .. */
+		if (strcmp(entry->d_name, ".") == 0 || strcmp(entry->d_name, "..") == 0)
+			continue;
+
+		if (snprintf(path, sizeof(path), "%s/%s", name, entry->d_name) >= sizeof(path)) {
+			fprintf(stderr, "Warning: path too long, skipping: %s/%s\n", name, entry->d_name);
+			continue;
+		}
+
+		/* Use lstat to check if it's a directory (portable) */
+		if (lstat(path, &entry_st) != 0) {
+			perror(path);
+			continue;
+		}
+
+		if (S_ISDIR(entry_st.st_mode)) { /* if it's a directory */
 			long startPos = lseek(ofd,0,1); /* remember where we are */
-			if (strcmp(entry->d_name, ".") == 0 || strcmp(entry->d_name, "..") == 0) {
-				continue;
-			}
 			if (verbose>1) {
 				for (i=0;i<level;i++) { fprintf(stdout, "  "); }
 				fprintf(stdout, "+ %s (directory)\n", entry->d_name);
@@ -255,8 +314,7 @@ int put_folder(char *name, int level)
 	return n;
 }
 
-int put_folder_entry(char *name, long startPos, int mtype, int level)
-{
+int put_folder_entry(char *name, long startPos, int mtype, int level) {
 	/* This function adds the special startFolder and endFolder entries that
 	   StuffIt uses to bracket the contents of a directory.
 
@@ -285,9 +343,15 @@ int put_folder_entry(char *name, long startPos, int mtype, int level)
 	long bs;
 
 	fpos1 = lseek(ofd,0,1); /* remember where we are (beginning of header) */
+	if (fpos1 < 0) {
+		fprintf(stderr, "Error seeking in archive: %s\n", strerror(errno));
+		return 0;
+	}
 	/* write empty header, will seek back and fill in later */
-	bzero(&fh,sizeof(fh));
-	write(ofd,&fh,sizeof(fh));
+	memset(&fh, 0, sizeof(fh));
+	if (safe_write(ofd, &fh, sizeof(fh), "folder header") < 0) {
+		return 0;
+	}
 
 	if (!(stat(name,&st)==0)) { /* get folder times */
 		perror(name);
@@ -302,17 +366,14 @@ int put_folder_entry(char *name, long startPos, int mtype, int level)
 	}
 	strncpy((char*)&fh.fName[1],fname,63); fh.fName[0] = min(strlen(fname),63);
 	ctime = st.st_ctime; /* ctime is really "time of last inode status change" */
+#ifdef HAVE_BIRTHTIME
 	ctime = st.st_birthtime; /* actual creation time is found in "birthtime" */
+#endif
 	mtime = st.st_mtime;
 	/* convert unix file time to mac time format */
-	time(&bs);
-	tp = localtime(&bs);
-	tdiff = TIMEDIFF + tp->tm_gmtoff;
-	if (tp->tm_isdst)
-		tdiff += 60 * 60;
+	tdiff = TIMEDIFF + get_timezone_offset();
 	cp4(ctime + tdiff,(char*)fh.cDate);
 	cp4(mtime + tdiff,(char*)fh.mDate);
-
 	fh.compRMethod = fh.compDMethod = mtype;
 	cp4(fpos1-startPos,(char*)fh.dLen);
 	cp4(fpos1-startPos,(char*)fh.cDLen);
@@ -320,29 +381,41 @@ int put_folder_entry(char *name, long startPos, int mtype, int level)
 	crc = updcrc(0,(unsigned char*)&fh,(sizeof(fh))-2);
 	cp2(crc,(char*)fh.hdrCRC);
 
-	fpos2 = lseek(ofd,0,1);		/* remember where we are (end of header) */
-	lseek(ofd,fpos1,0);				/* seek back to start of header */
-	write(ofd,&fh,sizeof(fh));		/* write back header */
+	fpos2 = lseek(ofd,0,SEEK_CUR);		/* remember where we are (end of header) */
+	if (fpos2 < 0 || lseek(ofd,fpos1,SEEK_SET) < 0) {
+		fprintf(stderr, "Error seeking in archive: %s\n", strerror(errno));
+		return 0;
+	}
+	if (safe_write(ofd, &fh, sizeof(fh), "folder header update") < 0) {
+		return 0;
+	}
 	if (mtype==endFolder) {
 		fh.compRMethod = fh.compDMethod = startFolder; /* fix up startFolder entry */
 		crc = updcrc(0,(unsigned char*)&fh,(sizeof(fh))-2);
 		cp2(crc,(char*)fh.hdrCRC);
-		lseek(ofd,startPos,0);		/* seek back to start of startFolder header */
-		write(ofd,&fh,sizeof(fh));	/* write back header */
+		if (lseek(ofd,startPos,SEEK_SET) < 0) {
+			fprintf(stderr, "Error seeking in archive: %s\n", strerror(errno));
+			return 0;
+		}
+		if (safe_write(ofd, &fh, sizeof(fh), "startFolder header update") < 0) {
+			return 0;
+		}
 	}
-	fpos2=lseek(ofd,fpos2,0);	/* seek forward file (back to end of header) */
+	if (lseek(ofd,fpos2,SEEK_SET) < 0) {
+		fprintf(stderr, "Error seeking in archive: %s\n", strerror(errno));
+		return 0;
+	}
 
 	return (fpos2 - fpos1);
 }
 
-int put_file(char *name, int level)
-{
+int put_file(char *name, int level) {
 	struct stat st;
 	struct infoHdr ih;
 	struct resHdr rh;
 	int i,n,fd;
 	long fpos1, fpos2;
-	char nbuf[256], *p;
+	char nbuf[PATH_MAX], *p;
 	int fork=0;
 	long tdiff;
 	size_t rlen, dlen;
@@ -350,38 +423,82 @@ int put_file(char *name, int level)
 	time_t ctime, mtime;
 	long bs;
 
-	fpos1 = lseek(ofd,0,1); /* remember where we are */
-	/* write empty header, will seek back and fill in later */
-	bzero(&fh,sizeof(fh));
-	write(ofd,&fh,sizeof(fh));
-
-	/* look for resource fork */
-	strcpy(nbuf,name);
-	strcat(nbuf,".rsrc");
-	if (!(stat(nbuf,&st)==0 && st.st_size)) { /* no .rsrc file */
-		strcpy(nbuf,name);
-		strcat(nbuf,"/..namedfork/rsrc"); /* try named fork */
-		if (!(stat(nbuf,&st)==0)) {
-			st.st_size = 0;
-		}
+	fpos1 = lseek(ofd,0,SEEK_CUR); /* remember where we are */
+	if (fpos1 < 0) {
+		fprintf(stderr, "Error seeking in archive: %s\n", strerror(errno));
+		return 0;
 	}
-	rlen = st.st_size;
-	if (st.st_size) {		/* resource fork exists */
-		dofork(nbuf,0);
-		cp4(st.st_size,(char*)fh.rLen);
-		cp4(clen,(char*)fh.cRLen);
-		cp2(crc,(char*)fh.rsrcCRC);
+	/* write empty header, will seek back and fill in later */
+	memset(&fh, 0, sizeof(fh));
+	if (safe_write(ofd, &fh, sizeof(fh), "file header") < 0) {
+		return 0;
+	}
+
+	/* look for resource fork - try multiple methods */
+	rlen = 0;
+
+	/* Method 1: Try AppleDouble sidecar file */
+	rlen = get_appledouble_rsrc_size(name);
+	if (rlen > 0) {
+		/* Write resource fork data and calculate CRC */
+		clen = read_appledouble_rsrc_with_crc(name, ofd, &crc, updcrc);
+		if (clen != rlen) {
+			fprintf(stderr, "Warning: resource fork size mismatch for %s\n", name);
+		}
+		cp4(rlen, (char*)fh.rLen);
+		cp4(clen, (char*)fh.cRLen);
+		cp2(crc, (char*)fh.rsrcCRC);
 		fh.compRMethod = noComp;
 		fork++;
 	}
-	strcpy(nbuf,name);
-	strcat(nbuf,".rsrc");
-	if (rmfiles) unlink(nbuf);	/* ignore errors */
+
+	/* Method 2: Try legacy .rsrc file */
+	if (rlen == 0) {
+		if (snprintf(nbuf, sizeof(nbuf), "%s.rsrc", name) >= sizeof(nbuf)) {
+			fprintf(stderr, "Error: path too long: %s.rsrc\n", name);
+			return 0;
+		}
+		if (stat(nbuf,&st)==0 && st.st_size) {
+			rlen = st.st_size;
+			dofork(nbuf,0);
+			cp4(st.st_size,(char*)fh.rLen);
+			cp4(clen,(char*)fh.cRLen);
+			cp2(crc,(char*)fh.rsrcCRC);
+			fh.compRMethod = noComp;
+			fork++;
+		}
+		if (rmfiles) unlink(nbuf);	/* ignore errors */
+	}
+
+#ifdef HAVE_NAMEDFORK
+	/* Method 3: Try macOS named fork */
+	if (rlen == 0) {
+		if (snprintf(nbuf, sizeof(nbuf), "%s/..namedfork/rsrc", name) >= sizeof(nbuf)) {
+			fprintf(stderr, "Error: path too long: %s/..namedfork/rsrc\n", name);
+			return 0;
+		}
+		if (stat(nbuf,&st)==0 && st.st_size) {
+			rlen = st.st_size;
+			dofork(nbuf,0);
+			cp4(st.st_size,(char*)fh.rLen);
+			cp4(clen,(char*)fh.cRLen);
+			cp2(crc,(char*)fh.rsrcCRC);
+			fh.compRMethod = noComp;
+			fork++;
+		}
+	}
+#endif
 
 	/* look for data fork */
-	strcpy(nbuf,name);
+	if (snprintf(nbuf, sizeof(nbuf), "%s", name) >= sizeof(nbuf)) {
+		fprintf(stderr, "Error: path too long: %s\n", name);
+		return 0;
+	}
 	if (stat(nbuf,&st)<0) {		/* first try plain name */
-		strcat(nbuf,".data");
+		if (snprintf(nbuf, sizeof(nbuf), "%s.data", name) >= sizeof(nbuf)) {
+			fprintf(stderr, "Error: path too long: %s.data\n", name);
+			return 0;
+		}
 		stat(nbuf,&st);
 	}
 	dlen = st.st_size;
@@ -400,9 +517,11 @@ int put_file(char *name, int level)
 	if (rmfiles) unlink(nbuf);	/* ignore errors */
 
 	/* look for .info file */
-	strcpy(nbuf,name);
-	strcat(nbuf,".info");
-	if ((fd=open(nbuf,0))>=0 && read(fd,&ih,sizeof(ih))==sizeof(ih)) {
+	if (snprintf(nbuf, sizeof(nbuf), "%s.info", name) >= sizeof(nbuf)) {
+		fprintf(stderr, "Error: path too long: %s.info\n", name);
+		return 0;
+	}
+	if ((fd=open(nbuf,O_RDONLY))>=0 && read(fd,&ih,sizeof(ih))==sizeof(ih)) {
 		/* %%% should we verify that .info file is exactly sizeof(ih)? */
 		strncpy((char*)fh.fName, ih.name,64);
 		strncpy((char*)fh.fType, ih.type, 4);
@@ -411,72 +530,91 @@ int put_file(char *name, int level)
 		strncpy((char*)fh.cDate, ih.ctime, 4);
 		strncpy((char*)fh.mDate, ih.mtime, 4);
 	}
-	else {	/* no info file, so try to use info from rsrc, else fake it */
+	else {	/* no info file, so try to use info from rsrc/AppleDouble, else fake it */
 		char *fName = basename(name);
+		AppleDoubleMetadata ad_meta;
+		int have_metadata = 0;
+
 		if (!fName) fName = name;
 		strncpy((char*)&fh.fName[1],fName,63); fh.fName[0] = min(strlen(fName),63);
 		/* default to LSC text file */
 		strncpy((char*)fh.fType, Type ? Type : "TEXT", 4);
 		strncpy((char*)fh.fCreator, Creator ? Creator : "KAHL", 4);
-		/* now try to read the actual type and creator from named fork */
-		strcpy(nbuf,name);
-		strcat(nbuf,"/..namedfork/rsrc");
-		if (rlen && (fd=open(nbuf,0))>=0 && read(fd,&rh,sizeof(rh))==sizeof(rh)) {
-			strncpy((char*)fh.fType, rh.type, 4);
-			strncpy((char*)fh.fCreator, rh.creator, 4);
-			strncpy((char*)fh.FndrFlags, rh.fdflags, 2);
+
+		/* Try AppleDouble metadata first */
+		if (read_appledouble_metadata(name, &ad_meta) == 0) {
+			strncpy((char*)fh.fType, ad_meta.type, 4);
+			strncpy((char*)fh.fCreator, ad_meta.creator, 4);
+			strncpy((char*)fh.FndrFlags, ad_meta.flags, 2);
+			have_metadata = 1;
 		}
+#ifdef HAVE_NAMEDFORK
+		/* Try reading from macOS named fork if AppleDouble didn't work */
+		if (!have_metadata) {
+			if (snprintf(nbuf, sizeof(nbuf), "%s/..namedfork/rsrc", name) >= sizeof(nbuf)) {
+				fprintf(stderr, "Warning: path too long for resource fork metadata: %s\n", name);
+			} else if (rlen && (fd=open(nbuf,O_RDONLY))>=0 && read(fd,&rh,sizeof(rh))==sizeof(rh)) {
+				strncpy((char*)fh.fType, rh.type, 4);
+				strncpy((char*)fh.fCreator, rh.creator, 4);
+				strncpy((char*)fh.FndrFlags, rh.fdflags, 2);
+				have_metadata = 1;
+			}
+		}
+#endif
 		ctime = st.st_ctime; /* ctime is really "time of last inode status change" */
+#ifdef HAVE_BIRTHTIME
 		ctime = st.st_birthtime; /* actual creation time is found in "birthtime" */
+#endif
 		mtime = st.st_mtime;
 		/* convert unix file time to mac time format */
-		time(&bs);
-		tp = localtime(&bs);
-		tdiff = TIMEDIFF + tp->tm_gmtoff;
-		if (tp->tm_isdst)
-			tdiff += 60 * 60;
+		tdiff = TIMEDIFF + get_timezone_offset();
 		cp4(ctime + tdiff, (char*)fh.cDate);
 		cp4(mtime + tdiff, (char*)fh.mDate);
 	}
-	close(fd);
-	strcpy(nbuf,name);
-	strcat(nbuf,".info");
-	if (rmfiles) unlink(nbuf);	/* ignore errors */
-
+	if (fd >= 0) close(fd);
+	if (snprintf(nbuf, sizeof(nbuf), "%s.info", name) < sizeof(nbuf)) {
+		if (rmfiles) unlink(nbuf);	/* ignore errors */
+	}
 	if (verbose) {
-		strncpy(nbuf,fh.fType, 4);
-		strncpy(nbuf+4,"/",1);
-		strncpy(nbuf+5,fh.fCreator,4);
-		strncpy(nbuf+9,"\0",1);
+		char typecreator[10];
+		snprintf(typecreator, sizeof(typecreator), "%.4s/%.4s", fh.fType, fh.fCreator);
 		if (verbose>1) { for (i=0;i<level;i++) { fprintf(stdout, "  "); } }
 		fprintf(stdout, "%s (%lld bytes) Data:%lld Rsrc:%lld [%s]\n",
-				name,(long long)dlen+rlen,(long long)dlen,(long long)rlen,nbuf);
+				name,(long long)dlen+rlen,(long long)dlen,(long long)rlen,typecreator);
 	}
 	crc = updcrc(0,(unsigned char*)&fh,(sizeof fh)-2);
 	cp2(crc,(char*)fh.hdrCRC);
 
-	fpos2 = lseek(ofd,0,1);		/* remember where we are */
-	lseek(ofd,fpos1,0);				/* seek back over file(s) and header */
-	write(ofd,&fh,sizeof fh);		/* write back header */
-	fpos2=lseek(ofd,fpos2,0);		/* seek forward file */
+	fpos2 = lseek(ofd,0,SEEK_CUR);		/* remember where we are */
+	if (fpos2 < 0 || lseek(ofd,fpos1,SEEK_SET) < 0) {
+		fprintf(stderr, "Error seeking in archive: %s\n", strerror(errno));
+		return 0;
+	}
+	if (safe_write(ofd, &fh, sizeof fh, "file header update") < 0) {
+		return 0;
+	}
+	if (lseek(ofd,fpos2,SEEK_SET) < 0) {
+		fprintf(stderr, "Error seeking in archive: %s\n", strerror(errno));
+		return 0;
+	}
 
 	return (fpos2 - fpos1);
 }
 
-int dofork(char *name, int convert)
-{
+int dofork(char *name, int convert) {
 	int fd, ufd, tfd;
 	size_t n;
 	char *p;
 	char *tempfilename = "sit+temp";
 
-	if ((fd=open(name,0))<0) {
+	if ((fd=open(name,O_RDONLY))<0) {
 		perror(name);
 		return 0;
 	}
 	if (convert) {	/* build conversion file */
 		if ((ufd=creat(tempfilename,0644))<0) {
 			perror(tempfilename);
+			close(fd);
 			return 0;
 		}
 	}
@@ -486,7 +624,11 @@ int dofork(char *name, int convert)
 		if (convert) {	/* convert '\n' to '\r' */
 			for (p=buf; p<&buf[n]; p++)
 				if (*p == '\n') *p = '\r';
-			write(ufd,buf,n);
+			if (safe_write(ufd, buf, n, "conversion temp file") < 0) {
+				close(fd);
+				if (convert) close(ufd);
+				return 0;
+			}
 		}
 		crc = updcrc(crc,(unsigned char*)buf,n);
 	}
@@ -494,13 +636,13 @@ int dofork(char *name, int convert)
 
 	if (convert) { /* reopen temp file */
 		close(ufd);
-		if ((tfd=open(tempfilename,0))<0) {
+		if ((tfd=open(tempfilename,O_RDONLY))<0) {
 			perror(tempfilename);
 			return 0;
 		}
 	}
 	else { /* reopen input file */
-		if ((tfd=open(name,0))<0) {
+		if ((tfd=open(name,O_RDONLY))<0) {
 			perror(name);
 			return 0;
 		}
@@ -508,7 +650,10 @@ int dofork(char *name, int convert)
 	/* write data to output archive */
 	clen = 0;
 	while ((n=read(tfd,buf,BUFSIZ))>0) {
-		write(ofd,buf,n);
+		if (safe_write(ofd, buf, n, "fork data") < 0) {
+			close(tfd);
+			return 0;
+		}
 		clen += n;
 	}
 	close(tfd);
@@ -516,14 +661,12 @@ int dofork(char *name, int convert)
 	return 0;
 }
 
-void cp2(unsigned int x, char *dest)
-{
+void cp2(uint16_t x, char *dest) {
 	dest[0] = x>>8;
 	dest[1] = x;
 }
 
-void cp4(unsigned long x, char *dest)
-{
+void cp4(uint32_t x, char *dest) {
 	dest[0] = x>>24;
 	dest[1] = x>>16;
 	dest[2] = x>>8;
